--- src/store/getCartShop.ts
@@ -1,4 +1,0 @@
-export const getcartShop = () => fetch('https://run.mocky.io/v3/d6e9a93f-9741-4494-b81e-637a8e9b8ddd')
-  .then(res => res.json())
-  .catch(error => console.error(error)
-)
